Move transcribe and translate tests from e2e to integ.
