+++ test/integ/run_integ.sh
@@ -67,16 +67,9 @@ echo "--------------------------------------------------------------------------
 echo "Setup test environment variables"
 
 export TEST_MEDIA_PATH="../test-media/"
-export TEST_IMAGE="sample-image.jpg"
-export TEST_VIDEO="sample-video.mp4"
-export TEST_AUDIO="sample-audio.m4a"
-export TEST_TEXT="sample-text.txt"
-export TEST_JSON="sample-data.json"
-export TEST_FACE_IMAGE="sample-face.jpg"
-export TEST_FACE_COLLECTION_ID="temporary_face_collection"
-
-# Retrieve exports from mie stack
-#export BUCKET_NAME=`aws cloudformation list-stack-resources --profile default --stack-name $MIE_STACK_NAME --region $REGION --output text --query 'StackResourceSummaries[?LogicalResourceId == \`Dataplane\`]'.PhysicalResourceId`
+export SAMPLE_TERMINOLOGY="sampleterminology"
+export TEST_PARALLEL_DATA_NAME="sample_parallel_data"
+export TEST_PARALLEL_DATA="sample_parallel_data.csv"
 
 echo "------------------------------------------------------------------------------"
 if [ "$1" = "" ]; then
