+++ test/integ/workflowapi/conftest.py
@@ -1,7 +1,7 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-
+import time
 import pytest
 import boto3
 import json
@@ -21,15 +21,11 @@ def testing_env_variables():
     try:
         test_env_vars = {
             'MEDIA_PATH': os.environ['TEST_MEDIA_PATH'],
-            'SAMPLE_IMAGE': os.environ['TEST_IMAGE'],
-            'SAMPLE_VIDEO': os.environ['TEST_VIDEO'],
-            'SAMPLE_AUDIO': os.environ['TEST_AUDIO'],
-            'SAMPLE_TEXT': os.environ['TEST_TEXT'],
-            'SAMPLE_JSON': os.environ['TEST_JSON'],
-            'SAMPLE_FACE_IMAGE': os.environ['TEST_FACE_IMAGE'],
+            'SAMPLE_TERMINOLOGY': os.environ['SAMPLE_TERMINOLOGY'],
+            'TEST_PARALLEL_DATA_NAME': os.environ['TEST_PARALLEL_DATA_NAME'],
+            'TEST_PARALLEL_DATA': os.environ['TEST_PARALLEL_DATA'],
             'REGION': os.environ['MIE_REGION'],
             'MIE_STACK_NAME': os.environ['MIE_STACK_NAME'],
-            'FACE_COLLECTION_ID': os.environ['TEST_FACE_COLLECTION_ID'],
             'ACCESS_KEY': os.environ['AWS_ACCESS_KEY_ID'],
             'SECRET_KEY': os.environ['AWS_SECRET_ACCESS_KEY']
             }
@@ -85,8 +81,9 @@ def stack_resources(testing_env_variables):
     return resources
 
 
-# API Class
+# Workflow API Class
 
+@pytest.mark.usefixtures("upload_media")
 class API:
     def __init__(self, stack_resources, testing_env_variables):
         self.env_vars = testing_env_variables
@@ -178,8 +175,126 @@ class API:
 
         return get_workflow_configuration_response
 
-# API Fixture
+    # Boto3 proxy methods
 
+    def get_terminology(self, body):
+        headers = {"Content-Type": "application/json"}
+        print("POST /service/translate/get_terminology")
+        get_terminology_response = requests.post(self.stack_resources["WorkflowApiEndpoint"]+'/service/translate/get_terminology', headers=headers, json=body, verify=True, auth=self.auth)
+
+        return get_terminology_response
+
+    def download_terminology(self, body):
+        headers = {"Content-Type": "application/json"}
+        print("POST /service/translate/download_terminology")
+        download_terminology_response = requests.post(self.stack_resources["WorkflowApiEndpoint"]+'/service/translate/download_terminology', headers=headers, json=body, verify=True, auth=self.auth)
+
+        return download_terminology_response
+
+    def list_terminologies(self):
+        print("GET /service/translate/list_terminologies")
+        list_terminologies_response = requests.get(self.stack_resources["WorkflowApiEndpoint"]+'/service/translate/list_terminologies', verify=True, auth=self.auth)
+
+        return list_terminologies_response
+
+    def delete_terminology(self, body):
+        headers = {"Content-Type": "application/json"}
+        print("POST /service/translate/delete_terminology")
+        delete_terminology_response = requests.post(self.stack_resources["WorkflowApiEndpoint"]+'/service/translate/delete_terminology', headers=headers, json=body, verify=True, auth=self.auth)
+
+        return delete_terminology_response
+
+    def create_terminology(self, body):
+        headers = {"Content-Type": "application/json"}
+        print("POST /service/translate/create_terminology")
+        create_terminology_response = requests.post(self.stack_resources["WorkflowApiEndpoint"]+'/service/translate/create_terminology', headers=headers, json=body, verify=True, auth=self.auth)
+
+        return create_terminology_response
+
+    def get_parallel_data(self, body):
+        headers = {"Content-Type": "application/json"}
+        print("POST /service/translate/get_parallel_data")
+        get_parallel_data_response = requests.post(self.stack_resources["WorkflowApiEndpoint"]+'/service/translate/get_parallel_data', headers=headers, json=body, verify=True, auth=self.auth)
+
+        return get_parallel_data_response
+
+    def list_parallel_data(self):
+        print("GET /service/translate/list_parallel_data")
+        list_parallel_data_response = requests.get(self.stack_resources["WorkflowApiEndpoint"]+'/service/translate/list_parallel_data', verify=True, auth=self.auth)
+
+        return list_parallel_data_response
+
+    def download_parallel_data(self, body):
+        headers = {"Content-Type": "application/json"}
+        print("POST /service/translate/download_parallel_data")
+        download_parallel_data_response = requests.post(self.stack_resources["WorkflowApiEndpoint"]+'/service/translate/download_parallel_data', headers=headers, json=body, verify=True, auth=self.auth)
+
+        return download_parallel_data_response
+
+    def delete_parallel_data(self, body):
+        headers = {"Content-Type": "application/json"}
+        print("POST /service/translate/delete_parallel_data")
+        delete_parallel_data_response = requests.post(self.stack_resources["WorkflowApiEndpoint"]+'/service/translate/delete_parallel_data', headers=headers, json=body, verify=True, auth=self.auth)
+
+        return delete_parallel_data_response
+
+    def create_parallel_data(self, body):
+        headers = {"Content-Type": "application/json"}
+        print("POST /service/translate/create_parallel_data")
+        create_parallel_data_response = requests.post(self.stack_resources["WorkflowApiEndpoint"]+'/service/translate/create_parallel_data', headers=headers, json=body, verify=True, auth=self.auth)
+
+        return create_parallel_data_response
+
+    def list_language_models(self):
+        headers = {"Content-Type": "application/json"}
+        print("GET /service/transcribe/list_language_models")
+        list_language_models_response = requests.get(self.stack_resources["WorkflowApiEndpoint"]+'/service/transcribe/list_language_models', headers=headers, verify=True, auth=self.auth)
+
+        return list_language_models_response
+
+    def describe_language_model(self, body):
+        headers = {"Content-Type": "application/json"}
+        print("POST /service/transcribe/describe_language_model")
+        describe_language_model_response = requests.post(self.stack_resources["WorkflowApiEndpoint"]+'/service/transcribe/describe_language_model', headers=headers, json=body, verify=True, auth=self.auth)
+
+        return describe_language_model_response
+
+
+@pytest.fixture(scope='session', autouse=True)
+def upload_media(testing_env_variables, stack_resources):
+    print('Uploading Test Media')
+    s3 = boto3.client('s3', region_name=testing_env_variables['REGION'])
+    # Upload test media files
+    s3.upload_file(testing_env_variables['MEDIA_PATH'] + testing_env_variables['TEST_PARALLEL_DATA'], stack_resources['DataplaneBucket'], testing_env_variables['TEST_PARALLEL_DATA'])
+    # Wait for fixture to go out of scope:
+    yield upload_media
+
+@pytest.fixture(scope='session')
+def terminology(workflow_api, stack_resources, testing_env_variables):
+    workflow_api = workflow_api()
+
+    create_terminology_body = {
+        "terminology_name": testing_env_variables['SAMPLE_TERMINOLOGY'],
+        "terminology_csv": "\"en\",\"es\"\n\"STEEN\",\"STEEN-replaced-by-terminology\""
+    }
+
+    create_terminology_request = workflow_api.create_terminology(
+        create_terminology_body)
+    assert create_terminology_request.status_code == 200
+
+    # wait for terminology to complete - it's synchronous but just in case
+    time.sleep(5)
+
+    yield create_terminology_body
+    delete_terminology_body = {
+        "terminology_name": testing_env_variables['SAMPLE_TERMINOLOGY']
+    }
+    delete_terminology_request = workflow_api.delete_terminology(
+        delete_terminology_body)
+    assert delete_terminology_request.status_code == 200
+
+
+# API Fixtures
 
 @pytest.fixture(scope='session')
 def workflow_api(stack_resources, testing_env_variables):
@@ -202,13 +317,12 @@ def api_schema():
 
     return schemata
 
-
 @pytest.fixture(scope='session')
 def operation_configs():
     return [
         {"Name":"video-test-async","Input":"Video", "Type":"Async", "Status":"OK"},
         {"Name": "video-test-sync", "Input": "Video", "Type": "Sync", "Status": "OK"}
-]
+    ]
 
 
 @pytest.fixture(scope='session')
@@ -218,14 +332,14 @@ def stage_configs():
          "ExecutedOperations": [], "Outputs": [], "Status": "OK"},
         {"Name": "video-async", "Input": "Video", "Operations": ["video-test-async"],
          "ExecutedOperations": [], "Outputs": [], "Status": "OK"}
-]
+    ]
 
 
 @pytest.fixture(scope='session')
 def workflow_configs():
     return [
         {"Name": "2-stage-wf", "Input": "Video", "Stages": ["video-sync", "video-async"], "Outputs": [], "Status": "OK"}
-]
+    ]
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -256,4 +370,5 @@ def cleanup(workflow_api, request, operation_configs, stage_configs, workflow_co
 
     request.addfinalizer(remove_operation_configs)
     request.addfinalizer(remove_stage_configs)
-    request.addfinalizer(remove_workflow_configs)
+    request.addfinalizer(remove_workflow_configs)\
+
