+++ test/integ/workflowapi/test_transcribe_proxy.py
@@ -15,28 +15,22 @@
 #
 ###############################################################################
 
-import urllib3
-import time
-import json
-import pytest
-import os
-
 def test_custom_language_model(workflow_api, testing_env_variables):
     workflow_api = workflow_api()
 
     # List custom language models.
 
-    list_custom_language_models_response = workflow_api.list_custom_language_models()
-    assert list_custom_language_models_response.status_code == 200
-    response = list_custom_language_models_response.json()
+    list_language_models_response = workflow_api.list_language_models()
+    assert list_language_models_response.status_code == 200
+    response = list_language_models_response.json()
     assert "Models" in response
 
     # Describe a custom language model if any exist.
 
-    if len(response["Models"] > 0):
+    if len(response["Models"]) > 0:
         model_name = response["Models"][0]["ModelName"]
         body = {'model_name': model_name}
-        describe_custom_language_model_response = workflow_api.describe_custom_language_model(body)
-        response = describe_custom_language_model_response.json()
+        describe_language_model_response = workflow_api.describe_language_model(body)
+        response = describe_language_model_response.json()
         assert "ModelName" in response["LanguageModel"]
         assert response["LanguageModel"]["ModelName"] == model_name
